# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campaign API",
        "description": "API documentation for Campaign management system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/news-sources": {
            "get": {
                "tags": [
                    "News Sources"
                ],
                "summary": "Get paginated news sources",
                "operationId": "e3a72259ae9232c551a4a9ab5655e427",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NewsSource"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "News Sources"
                ],
                "summary": "Create news source",
                "operationId": "555850445d7818cf14c40aa3ce69276e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewsSource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/NewsSource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/news-sources/{id}": {
            "get": {
                "tags": [
                    "News Sources"
                ],
                "summary": "Get news source by ID",
                "operationId": "b7f7b8cc39826dfed60c80e20ee8df33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/NewsSource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "News Sources"
                ],
                "summary": "Update news source",
                "operationId": "3e5b3689974aec730881c4329b948e24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewsSource"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "News Sources"
                ],
                "summary": "Delete news source",
                "operationId": "fb0840493dce84cb8fcae0549ff09878",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/news-sources/type/{type}": {
            "get": {
                "tags": [
                    "News Sources"
                ],
                "summary": "Get news sources by type",
                "operationId": "474d976cb9e218ad852291160e4dc4fc",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "website",
                                "rss",
                                "api"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NewsSource": {
                "description": "News source model",
                "properties": {
                    "id": {
                        "description": "News source ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Source name",
                        "type": "string",
                        "maxLength": 255,
                        "example": "BBC News"
                    },
                    "base_url": {
                        "description": "Base URL of the source",
                        "type": "string",
                        "example": "https://www.bbc.com"
                    },
                    "source_type": {
                        "description": "Type of source",
                        "type": "string",
                        "enum": [
                            "website",
                            "rss",
                            "api"
                        ],
                        "example": "rss",
                        "nullable": true
                    },
                    "crawl_interval_minutes": {
                        "description": "Crawl interval in minutes",
                        "type": "integer",
                        "example": 60
                    },
                    "is_active": {
                        "description": "Whether the source is active",
                        "type": "boolean",
                        "example": true
                    },
                    "last_crawled_at": {
                        "description": "Last crawl timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}